# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: PetSymphony CRM API
  description: REST API для управления клиентами, контактами и заметками
  version: 0.1.0
servers:
  - url: https://localhost/api
    description: Dev (FrankenPHP)
...
paths:
  /clients:
    get:
      tags: [clients]
      summary: Список клиентов
      description: Обычный пользователь видит только клиентов, которыми владеет; ROLE_MODERATOR и ROLE_ADMIN — всех. Пер-объектные эндпоинты отвечают 404 на чужого клиента.
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
            maximum: 100
        - name: search
          in: query
          description: Поиск по имени, email и телефону (подстрока, без регистра)
          schema:
            type: string
        - name: archived
          in: query
          description: true — включить архивных клиентов в выдачу
          schema:
            type: boolean
            default: false
        - name: tags
          in: query
          description: Имена тегов через запятую — фильтр по ИЛИ
          schema:
            type: string
            example: вокал,онлайн
        - name: instrument
          in: query
          description: id инструмента справочника — фильтр по инструменту
          schema:
            type: integer
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Client'
                  total:
                    type: integer
                  page:
                    type: integer
                  limit:
                    type: integer
    post:
      tags: [clients]
      summary: Создать клиента
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ClientInput'
      responses:
        '201':
          description: Создан
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Client'
        '422':
          $ref: '#/components/responses/ValidationError'
  /clients/export:
    get:
      tags: [clients]
      summary: Экспорт клиентов в CSV
      description: Те же фильтры, что у списка (search, archived, tags). UTF-8 с BOM, стриминг. Колонки — name, email, phone, comment, tags («; »-разделитель), created_at, archived_at.
      parameters:
        - name: search
          in: query
          schema:
            type: string
        - name: archived
          in: query
          schema:
            type: boolean
        - name: tags
          in: query
          schema:
            type: string
      responses:
        '200':
          description: CSV-файл
          content:
            text/csv:
              schema:
                type: string
  /clients/import:
    post:
      tags: [clients]
      summary: Импорт клиентов из CSV
      description: 'Колонки по заголовкам (рус/англ): name/имя (обязательно), email, phone, comment, tags (внутри ячейки — через «;»). Кодировки UTF-8 и Windows-1251. До 5000 строк, файл до 5 МБ, всё в одной транзакции; ошибочные строки не прерывают импорт и попадают в отчёт.'
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file]
              properties:
                file:
                  type: string
                  format: binary
                duplicates:
                  type: string
                  enum: [skip, update]
                  default: skip
                  description: Что делать с дублями по email
      responses:
        '200':
          description: Отчёт импорта
          content:
            application/json:
              schema:
                type: object
                properties:
                  created:
                    type: integer
                  updated:
                    type: integer
                  skipped:
                    type: integer
                  errors:
                    type: array
                    items:
                      type: object
                      properties:
                        line:
                          type: integer
                          description: Номер строки файла (0 — ошибка всего файла)
                        message:
                          type: string
        '422':
          $ref: '#/components/responses/ValidationError'
...
components:
  schemas:
    Client:
      type: object
      properties:
        id:
          type: integer
          readOnly: true
        name:
          type: string
        email:
          type: string
          format: email
          nullable: true
        phone:
          type: string
          nullable: true
        comment:
          type: string
          nullable: true
        tags:
          type: array
          readOnly: true
          items:
            type: string
          description: Нормализованные имена тегов (lower-case), отсортированы
        instruments:
          type: array
          readOnly: true
          items:
            type: object
            properties:
              id:
                type: integer
              name:
                type: string
          description: Инструменты ученика из справочника
        createdAt:
          type: string
          format: date-time
          readOnly: true
        updatedAt:
          type: string
          format: date-time
          readOnly: true
          nullable: true
        archivedAt:
          type: string
          format: date-time
          readOnly: true
          nullable: true
    ClientInput:
      type: object
      required: [name]
      properties:
        name:
          type: string
          minLength: 2
          maxLength: 180
        email:
          type: string
          format: email
          nullable: true
        phone:
          type: string
          maxLength: 32
          nullable: true
        comment:
          type: string
          maxLength: 10000
          nullable: true
        tags:
          type: array
          maxItems: 20
          items:
            type: string
            maxLength: 40
          description: Полный набор тегов клиента; несуществующие создаются автоматически
        instrumentIds:
          type: array
          maxItems: 20
          items:
            type: integer
          description: id инструментов справочника; неизвестные молча отбрасываются
...
    Error:
      description: Единый конверт ошибок API. message есть всегда; errors — карта «поле → сообщение» только у ошибок валидации, иначе null.
      type: object
      required: [message, errors]
      properties:
        message:
          type: string
        errors:
          type: object
          nullable: true
          additionalProperties:
            type: string
  responses:
...
    ValidationError:
      description: Ошибка валидации
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'